+++ dart_exporter/example_package/lib/example_package.dart
@@ -2,5 +2,3 @@
 ///
 /// More dartdocs go here.
 library example_package;
-
-export 'src/exports.dart_exporter.dart';