+++ dart_exporter/example_package/lib/src/exports.dart_exporter.dart
@@ -1,12 +1,7 @@
-// 
+// example_package
 //! AUTO GENERATE FILE, DONT MODIFY!!
 // ignored: 
 //.g.dart
 //.gen.dart
 //khoa.dart
-export 'package:example_package/example_package.dart';
-//ignored_file_extension [khoa.dart]:  export 'package:example_package/src/khoa.dart';
-//ignored_file_extension [.g.dart]:  export 'package:example_package/src/khoa.g.dart';
-//ignored_file_extension [.g.dart]:  export 'package:example_package/src/khoa.khoa.g.dart';
-export 'package:example_package/src/test.dart';
-//ignored_file_extension [.gen.dart]:  export 'package:example_package/src/khoa.gen.dart';
+export 'package:example_package/src/test.dart' show B,C;