+++ dart_exporter/example_package/lib/src/test.dart
@@ -1,0 +1,11 @@
+import 'package:dart_exporter_annotation/dart_exporter_annotation.dart';
+
+class A {}
+
+@Export()
+class B {}
+
+class _A {}
+
+@Export()
+class C {}