+++ dart_exporter_annotation/lib/dart_exporter_annotation.dart
@@ -3,8 +3,10 @@
 /// More dartdocs go here.
 library dart_exporter_annotation;
 
-const doNotExport = DoNotExport();
-
 class DoNotExport {
   const DoNotExport();
+}
+
+class Export {
+  const Export();
 }