+++ dart_exporter_annotation/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2.0.0
+- add ```Export``` annotation
+
 ## 1.0.0
 
 - Initial version.